PopUp feature
"The PopUp feature has been implemented in the ChatActivity. Now, when you enter a chat password and click the 'Далі' button, the password will be written to the file 'contactkeymap' as a value in the 'HashMap<String, String>'. The key will be the contact number and the value will be the chat password."